Setup ADC in adc module.
